backend/pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "Agent management platform for companies using AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "pygithub>=2.5.0",
    "pydantic-settings>=2.7.0",
    "greenlet>=3.1.0",
    "uvicorn>=0.34.0",
    "pyjwt>=2.10.0",
    "pwdlib[argon2]>=0.3.0",
    "itsdangerous>=2.2.0",
    "slowapi>=0.1.9",
    "email-validator>=2.2.0",
    "fastapi-mail>=1.6.0",
    "python-multipart>=0.0.18",
    "pyyaml>=6.0.0",
    "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atlas"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
